+++ services/dask-sidecar/src/simcore_service_dask_sidecar/settings.py
@@ -4,7 +4,7 @@ from typing import Any, Optional
 from models_library.basic_types import LogLevel
 from pydantic import Field, validator
 from settings_library.base import BaseCustomSettings
-from settings_library.logging_utils import MixinLoggingSettings
+from settings_library.utils_logging import MixinLoggingSettings
 
 
 class Settings(BaseCustomSettings, MixinLoggingSettings):
