+++ services/datcore-adapter/src/simcore_service_datcore_adapter/core/settings.py
@@ -4,8 +4,8 @@ from models_library.basic_types import BootModeEnum, BuildTargetEnum, LogLevel
 from pydantic import Field
 from pydantic.networks import AnyUrl
 from settings_library.base import BaseCustomSettings
-from settings_library.logging_utils import MixinLoggingSettings
 from settings_library.tracing import TracingSettings
+from settings_library.utils_logging import MixinLoggingSettings
 
 
 class PennsieveSettings(BaseCustomSettings):
