+++ services/storage/src/simcore_service_storage/cli.py
@@ -1,7 +1,7 @@
 import logging
 
 import typer
-from settings_library.cli_utils import create_settings_command
+from settings_library.utils_cli import create_settings_command
 
 from . import application
 from .settings import Settings
