+++ services/storage/src/simcore_service_storage/settings.py
@@ -3,10 +3,10 @@ from typing import List, Optional
 from pydantic import Field, PositiveInt, validator
 from settings_library.base import BaseCustomSettings
 from settings_library.basic_types import LogLevel, PortInt
-from settings_library.logging_utils import MixinLoggingSettings
 from settings_library.postgres import PostgresSettings
 from settings_library.s3 import S3Settings
 from settings_library.tracing import TracingSettings
+from settings_library.utils_logging import MixinLoggingSettings
 
 from .datcore_adapter.datcore_adapter_settings import DatcoreAdapterSettings
